requirements.txt
fastapi
uvicorn[standard]
sqlmodel
SQLAlchemy
psycopg2-binary
pydantic
python-multipart
passlib[bcrypt]
PyJWT
openpyxl
